+++ packages/firebase_auth/CHANGELOG.md
@@ -1,7 +1,19 @@
-## 0.8.1+6
+## 0.8.3+1
 
 * Bump the minimum flutter version to 1.2.0.
 * Add Template type parameter to `invokeMethod` calls.
+
+## 0.8.3
+
+* Make providerId 'const String' to use in 'case' statement.
+
+## 0.8.2+1
+
+* Fixed bug where `PhoneCodeAutoRetrievalTimeout` callback was never called.
+
+## 0.8.2
+
+* Fixed `linkWithCredential` on Android.
 
 ## 0.8.1+5